requirements.txt
flask
flask-cors
python-dotenv
pdfplumber
langchain-perplexity
streamlit
requests
openpyxl